requirements.txt
pandas
pyarrow
boto3
scikit-learn
numpy
pytest
transformers<4.46
torch>=2.3,<2.6 ; sys_platform == 'darwin' and platform_machine == 'arm64'
PyYAML
# Enhanced pipeline dependencies
wordcloud
matplotlib
seaborn